ensure MORPH_API_KEY is set with system check framework
