+++ polling_stations/apps/data_collection/management/commands/__init__.py
@@ -6,5 +6,4 @@ from data_collection.base_importers import (
     BaseCsvStationsCsvAddressesImporter,
     BaseShpStationsCsvAddressesImporter,
     BaseApiKmlStationsKmlDistrictsImporter,
-    BaseMorphApiImporter,
 )
